Type HelloWorld hooks dummy flow
