+++ react_on_rails/spec/dummy/client/app/startup/HelloWorldHooks.tsx
@@ -1,26 +1,31 @@
 // Super simple example of the simplest possible React component
 import React, { useState } from 'react';
-import PropTypes from 'prop-types';
+
 import css from '../components/HelloWorld.module.scss';
 
+import type { HelloWorldData } from './HelloWorld';
+
+type HelloWorldHooksProps = {
+  helloWorldData: HelloWorldData;
+};
+
 // TODO: make more like the HelloWorld.tsx
-function HelloWorldHooks({ helloWorldData }) {
+function HelloWorldHooks({ helloWorldData }: HelloWorldHooksProps) {
   const [name, setName] = useState(helloWorldData.name);
   return (
     <div>
       <h3 className={css.brightColor}>Hello, {name}!</h3>
       <p>
         Say hello to:
-        <input type="text" value={name} onChange={(e) => setName(e.target.value)} />
+        <input
+          type="text"
+          value={name}
+          onChange={(event: React.ChangeEvent<HTMLInputElement>) => setName(event.currentTarget.value)}
+        />
       </p>
     </div>
   );
 }
 
-HelloWorldHooks.propTypes = {
-  helloWorldData: PropTypes.shape({
-    name: PropTypes.string.isRequired,
-  }).isRequired,
-};
-
+export type { HelloWorldHooksProps };
 export default HelloWorldHooks;
