+++ react_on_rails/spec/dummy/client/app/startup/HelloWorldHooksContext.tsx
@@ -1,0 +1,45 @@
+// Example of using hooks when taking the props and railsContext
+// Note, you need the call the hooks API within the react component stateless function
+import React, { useState } from 'react';
+import type { RailsContext as RailsContextData } from 'react-on-rails/types';
+
+import css from '../components/HelloWorld.module.scss';
+import RailsContext from '../components/RailsContext';
+
+import type { HelloWorldData } from './HelloWorld';
+
+type HelloWorldHooksContextProps = {
+  helloWorldData: HelloWorldData;
+};
+
+type HelloWorldHooksContextRenderFunction = (
+  props: HelloWorldHooksContextProps,
+  railsContext: RailsContextData,
+) => React.ComponentType;
+
+// You could pass props here or use the closure
+const HelloWorldHooksContext: HelloWorldHooksContextRenderFunction = ({ helloWorldData }, railsContext) => {
+  const Result = () => {
+    const [name, setName] = useState(helloWorldData.name);
+    return (
+      <>
+        <h3 className={css.brightColor}>Hello, {name}!</h3>
+        <p>
+          Say hello to:
+          <input
+            type="text"
+            value={name}
+            onChange={(event: React.ChangeEvent<HTMLInputElement>) => setName(event.currentTarget.value)}
+          />
+        </p>
+        <p>Rails Context :</p>
+        <RailsContext {...{ railsContext }} />
+      </>
+    );
+  };
+
+  return Result;
+};
+
+export type { HelloWorldHooksContextProps, HelloWorldHooksContextRenderFunction };
+export default HelloWorldHooksContext;
